# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Merchant struct {
	MerchantID   int32            `json:"merchant_id"`
	MerchantName string           `json:"merchant_name"`
	Commission   float64          `json:"commission"`
	CreatedAt    pgtype.Timestamp `json:"created_at"`
	UpdatedAt    pgtype.Timestamp `json:"updated_at"`
}

type Transaction struct {
	TransactionID int32            `json:"transaction_id"`
	MerchantID    int32            `json:"merchant_id"`
	Amount        float64          `json:"amount"`
	Commission    float64          `json:"commission"`
	Fee           float64          `json:"fee"`
	CreatedAt     pgtype.Timestamp `json:"created_at"`
	UpdatedAt     pgtype.Timestamp `json:"updated_at"`
}